--- .eslintrc.json
@@ -1,19 +1,0 @@
-{
-  "extends": "next/core-web-vitals",
-  "plugins": ["prettier"],
-  "rules": {
-    "prettier/prettier": ["error", {
-      "endOfLine": "auto",
-      "columnWidth": 80,
-      "printWidth": 80,
-      "tabWidth": 4,
-      "useTabs": false,
-      "semi": true,
-      "singleQuote": true,
-      "trailingComma": "es5",
-      "bracketSpacing": true,
-      "jsxBracketSameLine": false,
-      "arrowParens": "always"
-    }]
-  }
-}
