--- .prettierrc.json
@@ -1,13 +1,0 @@
-{
-      "endOfLine": "auto",
-      "columnWidth": 80,
-      "printWidth": 80,
-      "tabWidth": 4,
-      "useTabs": false,
-      "semi": true,
-      "singleQuote": true,
-      "trailingComma": "es5",
-      "bracketSpacing": true,
-      "jsxBracketSameLine": false,
-      "arrowParens": "always"
-    }
