+++ components/uuuu/UuuCubeWrapper.tsx
@@ -1,18 +1,19 @@
 "use client";
-import MyonCube from "@/components/myon/Myoncube";
 import classes from "@/styles/3d.module.css";
 import { motion } from "framer-motion";
 import { useEffect, useState } from "react";
 import useSound from "use-sound";
+import UuuCube from "./UuuCube";
 
-const Page = () => {
+const UuuCubeWrapper = () => {
 	const [playbackRate] = useState(1);
 	const [isHovered, setIsHovered] = useState(false);
-	const [play, { stop, sound }] = useSound("/myon2.mp3", {
+	const [play, { stop, sound }] = useSound("/uuuu.mp3", {
 		volume: 0.25,
 		playbackRate,
 		interrupt: true,
 	});
+
 	const onHoverStart = () => {
 		setIsHovered(true);
 		play();
@@ -35,7 +36,6 @@ const Page = () => {
 			stop();
 		};
 	}, [stop]);
-
 	return (
 		<motion.main
 			initial={{ opacity: 0 }}
@@ -52,9 +52,10 @@ const Page = () => {
 				layout
 				className={classes.canvas}
 			>
-				<MyonCube playbackRate={playbackRate} isHovered={isHovered} />
+				<UuuCube isHovered={isHovered} />
 			</motion.div>
 		</motion.main>
 	);
 };
-export default Page;
+
+export default UuuCubeWrapper;
