+++ next.config.js
@@ -1,10 +1,10 @@
-const path = require('path');
+const path = require("path");
 /** @type {import('next').NextConfig} */
 const nextConfig = {
-    reactStrictMode: true,
-    sassOptions: {
-        includePaths: [path.join(__dirname, 'styles')],
-    },
+	reactStrictMode: true,
+	sassOptions: {
+		includePaths: [path.join(__dirname, "styles")],
+	},
 };
 
 module.exports = nextConfig;
