+++ styles/3d.module.css
@@ -1,37 +1,48 @@
 .canvas {
+  margin: auto;
+  height: 100vh;
+  width: 100vw;
+  image-rendering: -moz-crisp-edges;
+  image-rendering: -webkit-crisp-edges;
+  image-rendering: pixelated;
+  image-rendering: crisp-edges;
+}
+
+.festive {
+  background: linear-gradient(180deg, #01010c 0%, #000000 100%);
+  animation-delay: 2s;
+  animation: fade-in 2s ease-in-out;
+  background-repeat: no-repeat;
+}
+
+.title {
+  position: absolute;
+  z-index: 2;
+  top: 50%;
+  left: 50%;
+  transform: translateX(-50%);
+  font-size: 10em;
+  color: #fff;
+  font-weight: bold;
+}
+
+@media (max-width: 768px) {
+  .canvas {
     margin: auto;
-    height: calc(100vh - var(--header-height));
-    width: 100vw;
+    height: 50vh;
+    width: 100%;
     image-rendering: -moz-crisp-edges;
     image-rendering: -webkit-crisp-edges;
     image-rendering: pixelated;
     image-rendering: crisp-edges;
-}
-
-.festive {
-    background: linear-gradient(180deg, #01010c 0%, #000000 100%);
-    animation-delay: 2s;
-    animation: fade-in 2s ease-in-out;
-    background-repeat: no-repeat;
-}
-
-@media (max-width: 768px) {
-    .canvas {
-        margin: auto;
-        height: 50vh;
-        width: 100%;
-        image-rendering: -moz-crisp-edges;
-        image-rendering: -webkit-crisp-edges;
-        image-rendering: pixelated;
-        image-rendering: crisp-edges;
-    }
+  }
 }
 
 @keyframes fade-in {
-    0% {
-        background-size: 6000% 0%;
-    }
-    100% {
-        background-size: 100% 100%;
-    }
+  0% {
+    background-size: 6000% 0%;
+  }
+  100% {
+    background-size: 100% 100%;
+  }
 }
