+++ styles/globals.css
@@ -1,89 +1,89 @@
-@import url('https://cdn.jsdelivr.net/npm/galmuri@latest/dist/galmuri.css');
+@import url("https://cdn.jsdelivr.net/npm/galmuri@latest/dist/galmuri.css");
 
 :root {
-    --header-height: 29px;
-    --max-width: 1100px;
-    --border-radius: 12px;
-    --font-mono: ui-monospace, Menlo, Monaco, 'Cascadia Mono', 'Segoe UI Mono',
-        'Roboto Mono', 'Oxygen Mono', 'Ubuntu Monospace', 'Source Code Pro',
-        'Fira Mono', 'Droid Sans Mono', 'Courier New', monospace;
+  --header-height: 29px;
+  --max-width: 1100px;
+  --border-radius: 12px;
+  --font-mono: ui-monospace, Menlo, Monaco, "Cascadia Mono", "Segoe UI Mono",
+    "Roboto Mono", "Oxygen Mono", "Ubuntu Monospace", "Source Code Pro",
+    "Fira Mono", "Droid Sans Mono", "Courier New", monospace;
 
-    --foreground-rgb: 255, 255, 255;
-    --background-start-rgb: 0, 0, 0;
-    --background-end-rgb: 0, 0, 0;
+  --foreground-rgb: 255, 255, 255;
+  --background-start-rgb: 0, 0, 0;
+  --background-end-rgb: 0, 0, 0;
 
-    --primary-glow: radial-gradient(rgba(1, 65, 255, 0.4), rgba(1, 65, 255, 0));
-    --secondary-glow: linear-gradient(
-        to bottom right,
-        rgba(1, 65, 255, 0),
-        rgba(1, 65, 255, 0),
-        rgba(1, 65, 255, 0.3)
-    );
+  --primary-glow: radial-gradient(rgba(1, 65, 255, 0.4), rgba(1, 65, 255, 0));
+  --secondary-glow: linear-gradient(
+    to bottom right,
+    rgba(1, 65, 255, 0),
+    rgba(1, 65, 255, 0),
+    rgba(1, 65, 255, 0.3)
+  );
 }
 
 * {
-    box-sizing: border-box;
-    padding: 0;
-    margin: 0;
+  box-sizing: border-box;
+  padding: 0;
+  margin: 0;
 }
 
 html {
-    color-scheme: dark;
+  color-scheme: dark;
 }
 
 html,
 body {
-    font-family: Galmuri9, sans-serif;
-    font-size: 10px;
-    line-height: 20px;
-    max-width: 100vw;
-    overflow-x: hidden;
+  font-family: Galmuri9, sans-serif;
+  font-size: 10px;
+  line-height: 20px;
+  max-width: 100vw;
+  overflow-x: hidden;
 }
 
 body {
-    -webkit-font-smoothing: none;
-    font-smooth: never;
-    color: rgb(var(--foreground-rgb));
-    background: linear-gradient(
-            to bottom,
-            transparent,
-            rgb(var(--background-end-rgb))
-        )
-        rgb(var(--background-start-rgb));
+  -webkit-font-smoothing: none;
+  font-smooth: never;
+  color: rgb(var(--foreground-rgb));
+  background: linear-gradient(
+      to bottom,
+      transparent,
+      rgb(var(--background-end-rgb))
+    )
+    rgb(var(--background-start-rgb));
 
-    display: flex;
-    flex-direction: column;
-    max-width: inherit;
-    min-height: 100vh;
-    min-width: 100vw;
-    max-height: 100vh;
+  display: flex;
+  flex-direction: column;
+  max-width: inherit;
+  min-height: 100vh;
+  min-width: 100vw;
+  max-height: 100vh;
 }
 
 main {
-    position: relative;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    flex: 1;
+  position: relative;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  flex: 1;
 }
 
 a {
-    color: inherit;
-    text-decoration: none;
+  color: inherit;
+  text-decoration: none;
 }
 
 .scanlines {
-    position: fixed;
-    width: 100vw;
-    top: var(--header-height);
-    height: calc(100vh - var(--header-height));
-    pointer-events: none;
-    -webkit-user-select: none;
-    user-select: none;
-    background-size: 2px 3px;
-    background-image: radial-gradient(transparent 50%, black 50%);
-    z-index: 5;
+  position: fixed;
+  width: 100vw;
+  top: var(--header-height);
+  height: calc(100vh - var(--header-height));
+  pointer-events: none;
+  -webkit-user-select: none;
+  user-select: none;
+  background-size: 2px 3px;
+  background-image: radial-gradient(transparent 50%, black 50%);
+  z-index: 5;
 }
 
 h1,
@@ -116,16 +116,9 @@ textarea,
 select,
 button,
 svg {
-    margin: 0;
-    padding: 0;
-    border: 0;
-    vertical-align: middle;
-    line-height: 20px;
-}
-
-.active {
-    background: rgb(255, 136, 239);
-    color: #000;
-    line-height: 26px;
-    padding: 1px 30px;
+  margin: 0;
+  padding: 0;
+  border: 0;
+  vertical-align: middle;
+  line-height: 20px;
 }
