+++ styles/header.module.scss
@@ -1,65 +1,141 @@
 .header {
-    position: sticky;
-    color: #fff;
-    text-align: center;
-    display: flex;
-    gap: 10px;
-    border-bottom: 1px solid rgb(255, 136, 239);
-    min-height: var(--header-height);
-    justify-content: space-between;
-    z-index: 2000;
+  position: fixed;
+  width: 100%;
+  color: #fff;
+  text-align: center;
+  display: flex;
+  gap: 10px;
+  min-height: var(--header-height);
+  justify-content: space-around;
+  z-index: 2000;
+  top: 0;
+  padding: 10px;
+
+  &::before {
+    content: "";
+    position: absolute;
+    width: 100%;
+    height: 100%;
+    background: linear-gradient(
+      0deg,
+      rgba(0, 0, 0, 0) 0%,
+      rgba(0, 0, 0, 1) 100%
+    );
+    z-index: -1;
     top: 0;
+    left: 0;
+    right: 0;
+    bottom: 0;
+    opacity: 0;
+    backdrop-filter: blur(10px);
+  }
 
-    &>* {
-        margin-top: auto;
-        margin-bottom: auto;
-
-        &:last-of-type {
-            margin-right: 10px;
-        }
-
-        &>*:is(a) {
-            transition: padding 0.1s ease;
-            text-decoration: none;
-
-            &:hover {
-                text-decoration: underline;
-            }
-
-            &:active {
-                text-decoration: none;
-                color: rgb(255, 93, 182);
-            }
-        }
+  &:hover {
+    &::before {
+      opacity: 1;
     }
+    & > * {
+      & > *:is(a) {
+        & > span {
+          opacity: 1;
+          margin-top: 0;
+        }
+      }
+    }
+  }
+
+  & > * {
+    margin-top: auto;
+    margin-bottom: auto;
+
+    & > *:is(a) {
+      transition: all 0.1s ease;
+      text-decoration: none;
+      display: flex;
+      align-items: center;
+      flex-direction: column;
+      padding: 10px 15px;
+      filter: grayscale(100%);
+      overflow: hidden;
+      & > img {
+        outline: 1px solid rgba(0, 0, 0, 0);
+      }
+      & > span {
+        opacity: 0;
+        margin-top: -20px;
+        transition: opacity 0.1s ease, margin-top 0.1s ease;
+      }
+
+      &:hover,
+      &:focus {
+        text-decoration: underline;
+        & > img {
+          outline: 1px solid rgba(225, 0, 255, 0.5);
+        }
+        filter: grayscale(0%);
+      }
+
+      &:active,
+      &.active {
+        text-decoration: none;
+        & > img {
+          outline: 1px solid rgba(225, 0, 255, 1);
+        }
+        filter: grayscale(0%);
+      }
+    }
+  }
+
+  @media (max-width: 768px) {
+    top: unset;
+    bottom: 0;
+    &::before {
+      content: "";
+      position: absolute;
+      width: 100%;
+      height: 100%;
+      background: linear-gradient(
+        0deg,
+        rgba(0, 0, 0, 1) 0%,
+        rgba(0, 0, 0, 0.5) 100%
+      );
+      z-index: -1;
+      top: 0;
+      left: 0;
+      right: 0;
+      bottom: 0;
+      opacity: 0;
+      backdrop-filter: blur(10px);
+    }
+  }
 }
 
 .cube {
-    width: 100vw;
-    height: 28px;
+  width: 100vw;
+  height: 28px;
 }
 
 .marqueeContainer {
-    width: 100vw;
-    position: absolute;
-    top: 0;
-    left: 0;
-    z-index: 0;
+  width: 100vw;
+  position: absolute;
+  top: 0;
+  left: 0;
+  z-index: 0;
 }
 
 .scanlinesContainer {
-    position: relative;
+  position: relative;
 }
 
 .scanlines {
-    position: absolute;
-    width: 100vw;
-    top: 0;
-    height: var(--header-height);
-    pointer-events: none;
-    -webkit-user-select: none;
-    user-select: none;
-    background-size: 2px 3px;
-    background-image: radial-gradient(transparent 50%, black 50%);
-    z-index: 5;
+  position: absolute;
+  width: 100vw;
+  top: 0;
+  height: var(--header-height);
+  pointer-events: none;
+  -webkit-user-select: none;
+  user-select: none;
+  background-size: 2px 3px;
+  background-image: radial-gradient(transparent 50%, black 50%);
+  z-index: 5;
 }
